worked on the file upload/download functionality

+++ controllers/admin.js
@@ -1,6 +1,9 @@
 const { validationResult } = require('express-validator');
+const { file } = require('pdfkit');
 
 const Product = require('../models/product');
+
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -126,6 +129,7 @@ exports.postEditProduct = (req, res, next) => {
         }
         product.title = updatedTitle;
         if (image) {
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
         } 
         product.description = updatedDescription; 
@@ -160,8 +164,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ userId: req.user._id, _id: prodId })
-    .then(() => {
-        res.redirect('/admin/products');
-    }).catch(err => console.log(err));
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ userId: req.user._id, _id: prodId })
+        })
+        .then(() => {
+            res.redirect('/admin/products');
+        }).catch(err => console.log(err));
 };
